Task 2. Retrieve a list of the machines with OS 2
In this task, you need to obtain a list of all machines with the 'OS 2' operating system because these machines need an update. To get this information, you’ll run your first SQL query with a filter.

Select all the records from the machines table with a value of 'OS 2' in the operating_system column. Replace the value X with the correct string.